prevent compiler warnings.

+++ trunk/SDIF/test/utils.c
@@ -9,7 +9,7 @@ int writeFileHeader(SdifFileT *fileToWrite)
   size_t nbBytesWritten = 0;
   fprintf(stderr,"Write General Header (SdifFWriteGeneralHeader) \n");
   nbBytesWritten += SdifFWriteGeneralHeader(fileToWrite);
-  fprintf(stderr,"Number of bytes written: %d\n",nbBytesWritten);
+  fprintf(stderr,"Number of bytes written: %u\n",(unsigned int)nbBytesWritten);
   return nbBytesWritten; 
 }
 
@@ -18,7 +18,7 @@ int writeFileASCIIChunks(SdifFileT *fileToWrite)
   size_t nbBytesWritten = 0;
   fprintf(stderr,"Write ASCII Chunks (SdifFWriteAllASCIIChunks)\n");
   nbBytesWritten += SdifFWriteAllASCIIChunks(fileToWrite);
-  fprintf(stderr,"Number of bytes written: %d\n",nbBytesWritten);  
+  fprintf(stderr,"Number of bytes written: %u\n",(unsigned int)nbBytesWritten);  
   return nbBytesWritten;
 }
 
@@ -27,7 +27,7 @@ int writeFrameHeader(SdifFileT *fileToWrite)
   size_t nbBytesWritten = 0;
   fprintf(stderr,"Write frame header (SdifFWriteFrameHeader)\n");
   nbBytesWritten += SdifFWriteFrameHeader(fileToWrite);
-  fprintf(stderr,"Number of bytes written: %d\n",nbBytesWritten);
+  fprintf(stderr,"Number of bytes written: %u\n",(unsigned int)nbBytesWritten);
   return nbBytesWritten;
 }
 
@@ -37,9 +37,9 @@ int readFileHeader(SdifFileT *fileToRead)
 
   fprintf(stderr, "Read General Header (SdifFReadGeneralHeader) --> ");
   nbBytesRead += SdifFReadGeneralHeader(fileToRead);
-  fprintf(stderr, "%s\nNumber of bytes read: %d\n", 
+  fprintf(stderr, "%s\nNumber of bytes read: %u\n", 
 	  SdifSignatureToString(fileToRead->CurrSignature), 
-	  nbBytesRead);
+	  (unsigned int)nbBytesRead);
 
   return nbBytesRead;
 }
@@ -49,7 +49,7 @@ int readFileASCIIChunks(SdifFileT *fileToRead)
   size_t nbBytesRead = 0;
   fprintf(stderr,"Read ASCII Chunks (SdifFReadAllASCIIChunks) \n");
   nbBytesRead += SdifFReadAllASCIIChunks(fileToRead);
-  fprintf(stderr,"Number of bytes read: %d\n",nbBytesRead);
+  fprintf(stderr,"Number of bytes read: %u\n",(unsigned int)nbBytesRead);
   return nbBytesRead;   
 }
